Github_Bomb_Issues/requirements.txt
PyGithub
requests
os
getpass
pyjwt>=2.4.0 # not directly required, pinned by Snyk to avoid a vulnerability
